# authorization_service/auth-service/docs/openapi_auth_api_doc.yaml
openapi: 3.0.0
info:
  title: Auth API
  version: 1.0.0
paths:
...
  /admin/users/{user_id}/roles:
        get:
          tags:
            - Admin user permissions
          summary: Get user permissions (roles)
          parameters:
          - name: user_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          security:
            - Cookie: [access_token=]

          responses:
            '200':
              description: Ok
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/user-permissions-resp'
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
  /admin/users/{user_id}/roles/{role_id}:
        put:
          tags:
            - Admin user permissions
          summary: Assigne role on a user
          parameters:
          - name: user_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          - name: role_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          security:
            - Cookie: [access_token=]

          responses:
            '200':
              description: Ok
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/user-add-role-resp'
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
        delete:
          tags:
            - Admin user permissions
          summary: Remove role from a user
          parameters:
          - name: user_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          - name: role_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          security:
            - Cookie: [access_token=]

          responses:
            '200':
              description: Ok
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/user-add-role-resp'
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
  /admin/roles:
        get:
          tags:
            - Admin roles
          summary: Get list of roles
          security:
            - Cookie: [access_token=]

          responses:
            '200':
              description: Ok
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/roles-list-resp'
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
        post:
          tags:
            - Admin roles
          summary: Create a role
          requestBody:
            required: true
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    name:
                      type: string
          security:
            - Cookie: [access_token=]

          responses:
            '201':
              description: Ok
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/role-create-resp'
            '409':
              description: Role exists
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
  /admin/roles/{role_id}:
        get:
          tags:
            - Admin roles
          summary: Get role permissions
          parameters:
          - name: role_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          security:
            - Cookie: [access_token=]

          responses:
            '200':
              description: Ok
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/role-info-resp'
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
            '404':
              description: Role not found
        put:
          tags:
            - Admin roles
          summary: Update role permissions
          parameters:
          - name: role_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          requestBody:
            required: true
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    name:
                      type: string
                      example: "subscriber"
                    permissions:
                      type: array
                      example:
                        - "paid content viewer"

          security:
            - Cookie: [access_token=]

          responses:
            '200':
              description: Ok
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/role-info-resp'
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
            '404':
              description: Role not found
            '409':
              description: Role name aleady exists

        delete:
          tags:
            - Admin roles
          summary: Delete role
          parameters:
          - name: role_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
          security:
            - Cookie: [access_token=]

          responses:
            '200':
              description: Ok
            '404':
              description: Role not found
            '401':
              description: Unauthorised
              content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/unauthorised-resp'
components:
  schemas:
...
    unauthorised-resp:
      type: object
      properties:
        result:
          type: string
          example: 'unauthorised'
...
    user-permissions-resp:
      type: object
      properties:
        result:
          type: string
          example: 'success'
        data:
          type: string
          example: "{}"
        user_id:
          type: string
          example: 'b92ef010-5e4c-4fd0-99d6-41b6456272cd'
        roles:
          type: string
          example: "{uuid1: {role_name: subscriber, permissions: [ a, b, c]}}"
    user-add-role-resp:
      type: object
      properties:
        result:
          type: string
          example: 'success'
        user_id:
          type: string
          example: 'b92ef010-5e4c-4fd0-99d6-41b6456272cd'
        roles:
          type: array
          example: [{"id": "uuid1", "name": "subscriber"}, {"id": "uuid2", "name": "manager"}]
    roles-list-resp:
      type: object
      properties:
        result:
          type: string
          example: 'success'
        data:
          type: array
          example: [{"id": "uuid1", "name": "subscriber"}, {"id": "uuid2", "name": "manager"}]
    role-info-resp:
      type: object
      properties:
        result:
          type: string
          example: 'success'
        id:
          type: string
          example: 'uuid'
        name:
          type: string
          example: 'subscriber'
        permissions:
          type: array
          example: ["paid content viewer"]
    role-create-resp:
      type: object
      properties:
        result:
          type: string
          example: 'success'
        id:
          type: string
          example: 'uuid'
        name:
          type: string
          example: 'subscriber'